[TransferMatrixUser] fix source model creation
up to now, the transfermatrixuser class creates the source models for
both vector density types in the constructor. This leads to errors if a
source model is not supported for both version (e.g. subtraction only
supports dense vectors).
We thus create the source models depending on the density type.