[UDG] make level set data providable from outside
Optionally, level set images can now be provided from the outside and
do not have to be read from a nifti file. By this, we can directly pass
numpy arrays or matlab matrices to the driver without having to write
them to disk. A further modification should extend this to meshes for
fitted drivers.