[Fitted] let the user provide the grid
instead of reading the grid from file, the user can now provide the mesh
as a list of nodes and elements, along with labels and conductivity
values. The data is stored in the MEEGDriverData struct (along the
already existing UDG data). This way, we don't have to write the mesh to
a file when using something generate by other tools.