[MEG] use caching in integral assembler
The MEG integral assembler now supports caching of the assembled
integral vectors. If the integral assembler is bound to a list of coils
and projections and caching is desired, the integral vectors will be
assembled all at once.
Note that this will not (-epsilon) increase the computation time for a
single MEG forward computation, as the integrals have to be assembled
for each coil in any case.
As the MEG transfer matrix computation also uses the same integral
assembler subsequent calls to applyMEGTransferMatrix will directly use
the cached values.
The caching is enabled by default and can be disabled by setting
`cache.enable` to `false` in the meg parameter tree. As the computation
of the integral vectors takes some time, a verbose output can be enabled
by setting `verbose` to a positive integer in the meg parameter tree.