[LocalSub] use EntitySet instead of SubGrid
To speed up the computation of the localized subtraction, we introduce a special entity set to be used with the standard pdelab grid operator. The entity set only contains a subset of the elements of a certain gridview. In order to treat the intersection on the subset boundary properly, all intersection of the subset are wrapped in an intersection wrapper.