[Subtraction] add continuity type
Introduce an additional parameter to the subtraction source model
describing the continuity type of the discrete function space. For
a continuous function space, we do not have to integrate over the
skeleton and can thus skip it entirely in the assembly, which
speeds up the subtraction for cg. In Addition, we remove the
dependency on ConvectionDiffusionDGOperator, as we are only
interested in the lambda terms.