[PatchBasedVenant] add new source model for cg and dg
We add a variant of the venant approach that works for CG and DG.
Instead selecting the mesh vertices as monopoles, we select points
within elements surrounding the source. The elements are selected
using the element patch, that is already in use for the classical
venant and the localized subtraction. Within each element, monopoles
are selected based on the standard dune quadrature rule.