[Anisotropy] add anisotropic vtk functor
We add a vtk functor to export information about anisotropic
conductivity tensors. We use Eigen to compute the eigen values
and corresponding eigen vectors of the tensors. We scale the
eigen vectors by their eigen values (matrix is spd) and write
them to vtk. The vectors are sorted according to their eigen
values in increasing order.