[Anisotropy] change tensor representation
In the first implementation, tensors where given by providing one
tensor for each element. This commit changes the behavior by
expecting the user to provide one label per element and a list
of tensors (as many as there are different labels). The labels serve
as indices into the list of tensors. As we usually only consider
a single compartment for the anisotropy, this reduces the memory
consumption as we do not have to provide copies of the same tensor
for all isotropic elements but can use a single label for all of them
instead.