[TruncatedSpatialVenant] add truncated version
The standard spatial venant works only for CG and uses only those vertices
that are completely inside of the patch for the interpolation. By this
approach, the resulting source term is part of the ansatz space and has
a local support. With this commit, we add another spatial venant, which
uses all dofs associated with a patch. First of all, this enables the use
for both, CG and DG. For DG, we get a source term which is part of the
ansatz space. For CG, the source term is truncated at the patch boundary.