[UDG] make dipole and electrode compartment dynamic
The compartment of the dipoles and electrodes was selected at compile
time as the type of the grid function space had to be determined.
However, as we use a PowerGridFunctionSpace and all children have the
same type, we can extract the space dynamically so that we can select
the dipole and electrode compartment via the configuration.