[DipoleStatistics] add infrastructure
We add infrastructure to determine different statistics about a dipole. In
a first step, this is only the conductivity of the element the dipole is
in. A first implementation is provided for fitted meshes. This includes
tetrahedral, standard hexahedral and geometry adapted hexahedral meshes.
The configurartion for creating an object of this statistic is a subset
of the one for creating a MEEGDriver.