[PenaltyFluxWeighting] extract weight computation from LOP
For the DG, CutFEM and UDG methods, there were several places where the
weights for the penalty and average terms were computed. Besides being
just copy and pasted and thus error prone, it is also not very extensible
when one wants to try different weighting strategies.
This commit extracts the penalty and average weights from the local
operators and transfers them into an additional class, similar to the
edge norm provider.