[VolumeConductor] use always indirect mapping
In order to prepare for providing statistics about the volume conductor,
we remove the direct mapping to be able to access element labels.
Most of the meshes (but some legacy files) provide labeled elements and
we thus have a direct access to these labels.
For the legacy formats, this leads to a slight increase in time consumption
when accessing the tensor of an element, but compared to other tasks, this
should be negligible.
Besides being able to access the labels, this also gets rid of the type
erasure in the volume conductor, leading to way more readable code.