[MEEGDriver] add statistics method
We add a new method to the MEEGDriver interface, namely `statistics`.
The exact semantics of the method depends on the implementation, but
it should write out statistics about the driver to the datatree.
For the fitted driver, this currently means the volume and surface area
of the different discrete compartments.
For the unfitted driver, the semantics is similar as it writes out the
volume and surface area of the different domains.