Unify EEGForwardSolver
This unifies the fitted and unfitted EEGForwardSolver. Up to now, there
were two basically identical implementations, the
{Fitted,Unfitted}EEGForwardSolver. As they did basically the same thing, we
merge both implementations and parameterize them via the Solver and the
SourceModelFactory. While doing so, we move the element-search into the
Solver class and mark several variables as const that are nerver and also
should never be modified (mainly the VolumeConductor and the
ElementSearch).