[sparse vector] rewrite sparse vector implementation
sparse vector did not represent the hierarchical structure of
the function space, which makes it difficult to properly map them
to a flat representation.

We now create sparse vectors also as a hierarchical structure and add the
necessary tags to handle them in the PDELab ISTL backend.

(currently debug output is disabled)