fix bug in interpolation
If the dipole position happens to coincide with a mesh node the interpolation
introduces a NaN into the interpolated vector. If one then tries to
subtract the mean of the vector every entry gets replaced by NaN, and the
solution is destroyed. I opted to just replace NaN with zero at the dipole
position in this case. Since we are mostly interested in the potential
values at the surface this should not break any existing code.