Move element neighborhood map construction inside volume conductor storage creation
In the previous form, every iteration inside the parallel for loop of applyEEGTransfer_impl
and applyMEGTransfer_impl recomputes the element neighborhood map. For the
venant and EEG localized subtraction source model, this actually completeley
dominates the runtime. Performing the computation of the element neighborhood
map only once during driver creation, if necessary, can lead to an immense speedup