fix numerical flux test
By construction of the numerical flux local operator, the Raviart-Thomas
basis functions on boundary faces are assigned 0 as coefficient. When
comparing the corresponding interpolated grid function with a direct
gradient computation, we get an error in every element where boundary
Raviart-Thomas basis functions are supported. These are exactly the elements
possessing a boundary intersection. We circumvent this problem by only
comparing the gradients on elements without boundary intersections