refactor local subtraction code
Previously, the code for the continuous and the discontinuous local
subtraction computation did almost the same thing in two different
ways. The discontinuous code wrapped the patch in a new grid and used an
assembler from PDELab to assemble the right hand side, while the
continuous code used the element_patch_assembler from duneuro. This
commit refactors the code so that also in the discontinuous case, the
element patch assembler from duneuro is used. Furthermore, the code now
checks if the conductivity is isotropic before employing analytic
formulas for the right hand side.